Final version

+++ Candidates/HC_SR04.cpp
@@ -1,5 +1,4 @@
 #include "HC_SR04.h"
-#include <Arduino.h>
 
 HC_SR04::HC_SR04(int pinTrigger, int pinEcho){
   this -> pinTrigger = pinTrigger;
@@ -9,15 +8,19 @@ HC_SR04::HC_SR04(int pinTrigger, int pinEcho){
   pinMode(pinEcho, INPUT);
 }
 
-long HC_SR04:: getDistance(){
-  long duration;
+int HC_SR04:: getDistance(){
+  int duration;
   digitalWrite(pinTrigger, LOW);
+  
   delayMicroseconds(2);
   digitalWrite(pinTrigger, HIGH);
+
   delayMicroseconds(10); 
   digitalWrite(pinTrigger, LOW);
+   
   duration = pulseIn(pinEcho, HIGH);
   distance = (duration/2) / 29.1;
   return distance;
 }
+
